refactor: split up presenter and container for learn page
